+++ src/api/level/routes/01-custom-level.ts
@@ -1,0 +1,13 @@
+export default {
+  routes: [
+    {
+      method: 'POST',
+      path: '/levels/:id/unlock',
+      handler: 'level.unlock',
+      config: {
+        policies: [],
+        middlewares: [],
+      },
+    },
+  ],
+};