+++ tests/helpers/ctx-mock.ts
@@ -5,7 +5,8 @@ export type Ctx = {
   unauthorized: (msg: string, data?: any) => any
   notFound: (msg: string, data?: any) => any
   notImplemented: (msg: string, data?: any) => any
-  request?: any
+  request: { body?: any; [key: string]: any }
+  params: { [key: string]: any }
   query?: any
 }
 
@@ -16,6 +17,9 @@ export function mockCtx(user?: any): Ctx {
     badRequest: (msg: string, data?: any) => ({ status: 400, message: msg, data }),
     unauthorized: (msg: string, data?: any) => ({ status: 401, message: msg, data }),
     notFound: (msg: string, data?: any) => ({ status: 404, message: msg, data }),
-    notImplemented: (msg: string, data?: any) => ({ status: 501, message: msg, data })
+    notImplemented: (msg: string, data?: any) => ({ status: 501, message: msg, data }),
+    request: {},
+    params: {},
+    query: {}
   }
 }